requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9
matplotlib==3.8.4
seaborn==0.13.2
wordcloud==1.9.3
imblearn==0.0
imbalanced-learn==0.12.0
lightgbm==4.3.0
scikit-learn==1.4.2
joblib==1.3.2
nltk==3.8.1
pandas==2.2.2
numpy==1.26.4
tensorflow==2.16.1
keras==3.3.3
streamlit==1.34.0
requests==2.31.0
PyJWT==2.8.0

# For testing
pytest==7.4.4
httpx==0.27.0
